restructure the genere project by adding routes folder
