+++ building_Api3/express/server.js
@@ -1,0 +1,17 @@
+const express = require("express");
+const app = express();
+
+const genres = require('../routes/route')
+
+app.use(express.json());
+app.use('',genres)
+
+
+
+app.get("/", (req, res) => {
+  res.send("Hello World");
+});
+
+app.listen(3000, () => {
+  console.log("server is running on port 3000");
+});
